requirements.txt
Flask
pyrebase4
langchain
